+++ code/database/gui.py
@@ -5,55 +5,40 @@ from ttkthemes import ThemedTk
 import pandas as pd
 from pandasql import sqldf
 from pandastable import Table
+import functions as func
 
 
-#Read the csv into a panda df
-db = pd.read_csv("code/database/organism_database_20240428124116.csv")
+def main(db=func.read_csv()) -> None:
+    """"Main function for the GUI"""
+    #Create the main window
+    #root = tk.Tk()
+    root = ThemedTk(theme="arc")
+    root.title("Evolution Simulation")
 
-def querry_input():
-    """"Takes sql statement from user and returns the querried pandadf"""
-    user_querry = entry.get()
-    output = pysqldf(user_querry)
-    update_table(output)
+    #Define frame globally
+    frame = ttk.Frame(root)
+    frame.pack(fill='both', expand=True)
 
-def new_query():
-    """"Allows user to make a new querry"""
-    entry.delete(0, tk.END) #Clear entry widget
-    update_table(db)
+    #entry widget for user input
+    entry = ttk.Entry(root, width=40)
+    entry.pack(pady=10)
 
-def update_table(dataframe):
-    """"Update the table with new data"""
-    table.model.df = dataframe
-    table.redraw()
+    #submit button
+    submit_button = ttk.Button(root, text="Submit", command=func.querry_input)
+    submit_button.pack(pady=5)
 
-def pysqldf(query):
-    """Wrapper function for sqldf"""
-    return sqldf(query, globals())
+    #"restart" button
+    new_query_button = ttk.Button(root, text="restart", command=func.new_query)
+    new_query_button.pack(pady=5)
 
-#Create the main window
-#root = tk.Tk()
-root = ThemedTk(theme = "arc")
-root.title("Evolution Simulation")
+    #initial table display
+    table = Table(frame, dataframe=db)
+    table.setTheme("dark")
+    table.show()
 
-#Define frame globally
-frame = ttk.Frame(root)
-frame.pack(fill='both', expand=True)
+    root.mainloop()
 
-#entry widget for user input
-entry = ttk.Entry(root, width=40)
-entry.pack(pady=10)
 
-#submit button
-submit_button = ttk.Button(root, text="Submit", command=querry_input)
-submit_button.pack(pady=5)
-
-#"restart" button
-new_query_button = ttk.Button(root, text="restart", command=new_query)
-new_query_button.pack(pady=5)
-
-#initial table display
-table = Table(frame, dataframe=db)
-table.setTheme("dark")
-table.show()
-
-root.mainloop()
+if __name__ == "__main__":
+    db = func.read_csv()
+    main()
